Rename: Budget DTO 폴더 이동
기존에 분리되어 있었는데 dto로 폴더 생성후 data로 옮김
