+++ src/main/java/com/ssafy/tott/budget/controller/BudgetController.java
@@ -2,8 +2,8 @@ package com.ssafy.tott.budget.controller;
 
 import com.ssafy.tott.auth.annotation.Authenticated;
 import com.ssafy.tott.auth.vo.AuthMember;
-import com.ssafy.tott.budget.data.requset.BudgetsUpdateRequest;
-import com.ssafy.tott.budget.data.response.BudgetsResponse;
+import com.ssafy.tott.budget.data.dto.requset.BudgetsUpdateRequest;
+import com.ssafy.tott.budget.data.dto.response.BudgetsResponse;
 import com.ssafy.tott.budget.service.BudgetService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
